Updated all DTOs and service interfaces

+++ CinemaHikes/src/CinemaHikes.Application/Interfaces/Services/IReviewService.cs
@@ -5,5 +5,5 @@ namespace CinemaHikes.Application.Interfaces.Services;
 public interface IReviewService
 {
     Task<List<ReviewDto>> GetByMovieIdAsync(int movieId, CancellationToken ct);
-    Task CreateAsync(int userId, int movieId, string text, short rating, CancellationToken ct);
+    Task CreateAsync(int userId, int movieId, string text, double rating, CancellationToken ct);
 }
